--- backend/api/README.md
@@ -1,1 +1,0 @@
-This folder will contain the project code for the user-facing aspects of the project.
