--- backend/api/user.py
@@ -1,14 +1,0 @@
-#Front-End Facing Endpoints
-
-# To Do:
-# Endpoint Implements: All to pull from pre-existing data and display, not to create
-# 1. Retrieve Log
-# 2. Retrieve Temperature data
-# 3. Retrieve Moisture Data
-# 4. Retrieve MicroController Statuses
-
-# 5. Run Test Suite, display result
-# 6. Post whether or not black ice was recorded, IE send to system that black ice was recorded
-# 7. Github commit/issue tracking logs displayed on app
-
-
