+++ backend/classes/rasp_piClass.py
@@ -1,0 +1,4 @@
+from deviceClass import device
+
+class Pi(device):
+    sensor_readings: dict #{Temperature, Moisture, CPU_Temperature}
