+++ backend/settings/private.py
@@ -1,0 +1,5 @@
+class private:
+    sender = "example@email.com"
+    reciever ="example@email.com"
+    sender_password = "password"
+    host= "smtp.email.com"
